+++ src/fields/Checkboxes.php
@@ -318,6 +318,11 @@ class Checkboxes extends OptionsField
     // Protected Methods
     // =========================================================================
 
+    protected function supportedDefaults(): array
+    {
+        return ['layout'];
+    }
+
     protected function defineFieldSlotTag(string $key, RenderContext $context): ?SlotTag
     {
         $form = $context->form;